+++ src/lib/rt/src/rt_rtt_os.c
@@ -1536,7 +1536,7 @@ int     rtt_replace_env( char *str, char *newstr)
         char    *t;
         char    *u;
         int     symbolmode;
-        int     size;
+        long int size;
         char    *value;
         char    symbol[80];
         char    lower_symbol[80];
@@ -1558,7 +1558,7 @@ int     rtt_replace_env( char *str, char *newstr)
           else if (symbolmode && (*s == '/' || *s == '.'))
           {
             /* End of potential symbol */
-            size = (int) s - (int) u;
+            size = (long int) s - (long int) u;
             strncpy( symbol, u, size);
             symbol[size] = 0;
             cdh_ToLower( lower_symbol, symbol);
