+++ wb/lib/wb/src/wb_goenm14.cpp
@@ -106,7 +106,7 @@ int goen_create_nodetype_m14(
     pwr_tClassId	cid,
     ldh_tSesContext	ldhses,
     flow_tCtx		ctx,
-    unsigned long 	*mask,
+    unsigned int 	*mask,
     unsigned long	subwindowmark,
     unsigned long	node_width,
     flow_tNodeClass	*node_class,
@@ -118,9 +118,9 @@ int goen_create_nodetype_m14(
   int		inputpoints, outputpoints;
   unsigned long	inpointmask;
   unsigned long	outpointmask;
-  unsigned long	*inmask_pointer;
-  unsigned long	*outmask_pointer;
-  unsigned long	*invertmask_pointer;
+  unsigned int	*inmask_pointer;
+  unsigned int	*outmask_pointer;
+  unsigned int	*invertmask_pointer;
   ldh_sParDef 	*bodydef;
   int 		rows;
   char		*parvalue;
@@ -351,7 +351,7 @@ int goen_create_nodetype_m14(
 *
 **************************************************************************/
 int goen_get_point_info_m14( WGre *grectx, pwr_sGraphPlcNode *graphbody, 
-			    unsigned long point, unsigned long *mask, 
+			    unsigned long point, unsigned int *mask, 
 			    unsigned long node_width, goen_conpoint_type *info_pointer, 
 			    vldh_t_node node)
 {
@@ -441,7 +441,7 @@ int goen_get_point_info_m14( WGre *grectx, pwr_sGraphPlcNode *graphbody,
 **************************************************************************/
 int	goen_get_parameter_m14( pwr_sGraphPlcNode *graphbody, pwr_tClassId cid, 
 			       ldh_tSesContext ldhses, unsigned long con_point, 
-			       unsigned long *mask, unsigned long *par_type, 
+			       unsigned int *mask, unsigned long *par_type, 
 			       unsigned long *par_inverted, unsigned long *par_index)
 {
 
@@ -449,7 +449,7 @@ int	goen_get_parameter_m14( pwr_sGraphPlcNode *graphbody, pwr_tClassId cid,
 	unsigned long	conpointcount;
 	unsigned long	pointmask;
 	int		i, input_found, output_found;
-	unsigned long	*invertmask_pointer;
+	unsigned int	*invertmask_pointer;
 	int		rows;
 	ldh_sParDef 	*bodydef;
 
@@ -531,7 +531,7 @@ int	goen_get_parameter_m14( pwr_sGraphPlcNode *graphbody, pwr_tClassId cid,
 *
 **************************************************************************/
 int goen_get_location_point_m14( WGre *grectx, pwr_sGraphPlcNode *graphbody, 
-				unsigned long *mask, unsigned long node_width, 
+				unsigned int *mask, unsigned long node_width, 
 				goen_point_type *info_pointer, vldh_t_node node)
 {
    info_pointer->y = 0;
